Add alternative ( and better) solution for global calibrating the model
Once accepted the idea that a  different calibration has model implications and therefore can be treated as different model, then, w.r.t. the first one, this solution it's more simple, it exploits QL and creates just 2 new objects. ( but leaves a bit less control on the inner error calibration)